Fixed the issues due to the latest versions of R and testthat package
The use of version 2 of the testthat package has been made explicit in the DESCRIPTION file and fixed the crashes caused by the new behaviour introduced in version 3.0.2. Additionally, the form in which attributes are accessed in an array of the GenericPlot class is corrected due to the new R version. In addition, the documentation examples have been disabled until checking the proper operation of the package with the latest R version.